requirements.txt
python-dotenv
langchain 
openai
langchain-experimental
pandas
tabulate
streamlit